requirements.txt
dnspython==2.0.0
discord.py==1.5.0
pymongo==3.11.0